% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatus.R
\name{getStatus}
\alias{getStatus}
\title{Retrieve the Current System Status of the Kraken Exchange}
\usage{
getStatus(data = "both")
}
\arguments{
\item{data}{A character string specifying the type of data to return.
Use "status" for the system status, "timestamp" for
the timestamp, or "both" to retrieve both status and timestamp.
The default is "both".}
}
\value{
A character string if \code{data = "status"} or \code{data = "timestamp"}.
A named vector if \code{data = "both"}.
}
\description{
This function fetches the current system status from the Kraken API,
including both the operational status and the timestamp.
}
\examples{
getStatus("both")
getStatus("status")
getStatus("timestamp")
}
